Add `--precise` flag for upgrade
Fixes #177. If this flag is provided, all dependencies upgraded will be blindly
upgraded to the version specified.

Also, there has been substantial refactoring. The new `Manifests` and `Dependencies` structs now hold the logic for upgrade. This allows us to
commonise the cases of a single upgrade and full workspace upgrade quite
considerably. This should also make it fairly simple to extend to permit
providing a list of packages to confine upgrades to.

Also does some tidying up for the new `--precise` option. Most
significantly, we now have a test for this functionality.