Add '--to-lockfile' flag to cargo-upgrade
When this flag is present, all dependencies in 'Cargo.toml'
are upgraded to the locked version as recorded in 'Cargo.lock'.

Multiple manifests in a workspace may request the same dependency
at different versions. The version of each dependency is
updated to the corresponding semver compatible version in the
lockfile.