feat: --skip-compatible to avoid churn (#360)
This is a re-up of the abandoned #219 (a non-trivial merge) with the review comments addressed.

Original description:

Add a `--skip-compatible` flag to cargo upgrade

This flag will make `cargo upgrade` ignore upgrades where the old
version is semver compatible with the new one. This is useful in cases
where you don't want to churn the Cargo.toml files in the whole project
knowing that the lockfile is already forcing the versions to be up to
date.

To test that I had to cheat a little bit and hardcode two different fake
dependencies in `get_latest_dependency` which is unfortunate but
necessary as there's no way currently to mock answers from crates.io.