Update help output in readme file (#378)
* cargo-add: Explain values for --upgrade option in help

This patch adds an explanation of the possible values for the --upgrade
option of the cargo-add binary to its help output.  There was a similar
text in the old help output but it was removed during the migration to
structopt (f9b33f41080175451617ccd138ea5d6e1c360375).

* Update help output in readme file

This patch updates the help output of the cargo add, cargo rm and cargo
upgrade commands in the readme file as some new options (--offline,
--skip-compatible) were missing.

As the format of the help output has changed since the last update of
the readme, the complete output is replaced.