Remove `anyhow` from the library
As @yaahc pointed out, if we immediately erase the error type by
converting it to `anyhow::Error` then there's not much reason to have a
library error type.

This commit puts _all_ errors generated in the library into the Error
enum and removes use of `anyhow` from the library completely.

Note that `anyhow` is still used in the 3 binaries.