fix(add): Lookup dep versions *after* updating
I got a report of `cargo add inquire` not adding the latest version.
Each of us was getting a different version.  I suspect this is a bug in
the crates-index logic and want to port to `crates-index` to reduce the
chance of problems.

However, while looking at that, I found this problem.
`args.parse_dependencies` looks in the registry but we call it
**before** updating it.  This can lead us to reading stale data that, if
undid, the next call would get the latest.  So swapping the order of
operations to make sure we are reading the latest version of the index.