refactor: Pun index logic to crates-index (#484)
This was inspired by running into errors where `cargo add inquire` was
not adding the latest version.  We've seen similar problems during the
development of `crates-index`, so hoping what they've done fixes this.

Even if it doesn't solve this problem, hopefully by nature of using a
shared crate means it will get more use and more eyes to make it more
reliable in the long term.

Handling retrying the index is annoying.  I opened
https://github.com/frewsxcv/rust-crates-index/issues/62

Note: in the next version of crates-index
- Opening an index will implicitly create it, so we won't be able to
  differentiate in our message whether we are initializing or updating.
  See https://github.com/frewsxcv/rust-crates-index/issues/61
- We will be able to open an Index by URL, which will allow us to remove
  more logic