refactor: Exclusively switch to LocalManifest (#493)
For local content.  This is prep for fixing some bugs with workspaces
where `Dependency` has a path relative from the wrong location.  My plan
to fix it so for `Depednency` to have an absolute path.  Creating and
serializing `Dependency` will then require knowing what the path is for
the relevant manifest so we can handle converting the path.