fix(add): Form correct path dependencies
We originally treated `--path` as relative to each crate being added
instead of relative to the current working directory.  This caused users
confusion and doesn't work with tab completion.

Now, we'll treat the relative path as if its relative to the current
working directory, adapting it for each crate root.

I went back and forth on how to handle the tests.  Existing tests rely
on reusable fixtures, backed by files with functions to set them up.  I
considered updating the functions to produce fixtures needed for the new
local file requirements (paths exist) or to create new functions.  In
the end, I went with customizing it per function.  This made it so
assumptions in the test are contained within the test (e.g. dependency
name) rather than being split in separate areas.  It also made it easy
to customize the behavior, like adding version metadata for one test but

Mac and Windows also added difficulty in updating tests and required all
test inputs to be canonicalized.

Fixes #477
Fixes #453