list available features when adding a dependency through cargo-add (#538)
This one is to address #529 and does precisely that. One decision I made that the reviewer might want to consider is that I didn't add this functionality when using cargo-add to add specific versioned crates. That was mostly because that currently doesn't have to hit either the file system or the registry and I didn't want to add that there and basically duplicate all the effort, nor did I want to try and refactor the whole thing in this PR so for now I just left it. If this is something people would like to see I'm willing to work on that. Otherwise it's a relatively small PR, mostly just connecting the dots on work that was already done, so hopefully it will be an easy merge.

I also still have a draft PR (#528) that is related to this, and will at the least be easier to implement if this gets pulled. This does already do a long way to addressing the issue I wanted to address with that PR so if this gets pulled and nobody tells me that they would really like to see that one too, I might just close that. We'll see.

Over all I think it's a really nice QoL improvement and I hope you all like it :) Feedback is welcome as always.