test: Port set-version tests to trycmd
This is a test run for porting us from assert_cmd / assert_fs to trycmd.
The primary reason for this is to make it easier to update
expected output as I iterate on commands to merge into cargo.

This also has other benefits, like filesystem fixtures living on disk
for ease of porting reproductin case to test case, manual
experimentation, and for diffing inputs/outputs