fix(add): Don't report version for local paths
When its local, the only time the version matters is for end users, when
the crate is published.  This also helps clarify our workspace behavior.

I'd prefer to report the relevant path but we track the absolute path at
this point and make it relevant when we apply it to a manifest.