fix(add)!: Re-adding appends features
Take
```console
$ cargo add clap --features derive
$ cargo add clap --features unicode
```

Before, this would result in
```toml
clap = { version = "3.0.13", features = ["unicode"] }
```

Now, it results in
```toml
clap = { version = "3.0.13", features = ["derive", "unicode"] }
```

This is part of a larger effort to make `cargo add` "append-only".  We
should only modify fields users explicitly mention and, when we do, we
should append-or-replace (depending on the data type).