feat(add): Add `-F` short flag
This is serving as a "living" proposal for `-F` to be added to cargo
generally.  Before, we've rejected this because we wanted to mirror
cargo.

`-F` was chosen because `-f`'s de facto meaning is `--force`.  Even if
we don't have a `--force` flag, a cargo command could in the future and
we don't want conflicts.

Fixes #450