fix(add): Prefer copying existing dep requirements
When adding a dependency to multiple tables, most likely the user
intends to use the same version across all of them.  We should prefer
the existing version requirement, registry, and path.

This was inspired by a internals post where a user wanted to ellide the
version in dev/build dependencies when it also exists in the regular
dependencies.

See https://internals.rust-lang.org/t/allow-omission-of-version-for-dev-and-build-dependencies-that-are-also-regular-dependencies/16077