refactor: Shift no-dep-with-git/path error down a level
Now we pass into the op a crate name of `None` when there are no
positional arguments.  This is sort-of like how install is structured
(except the `None` handling is purely within the op).  Next, we'll just
need to look up a crate within the source and we're done.

I needed to move git/path handling earlier so we could use the
discovered name to find the `old_dep`.  This ended up simplifying the
code somewhat.

As a side benefit, the error message for `cargo add` (no args) is
clearer in what is required (once we can infer crate names).

Overall, there are a lot of ordering issues to deal with
- feature parsing needs the name of all crates (worked around by
  disallowing `dep/feature` when inferring).  We could move this later
  but we'd need to maintain the global nature of it to error when a
  qualified dependenyc is unused.
- getting the old depednency needs the dependency name which we
  will sometimes infer from the source
- getting the old dependency needs the dependency name normalized
  by the source (we just ignore that for the registry case)
- looking up the crate in the registry might need the source
  in the old depednency like version or registry (using cargo-add for
  modifying existing features)
- not all cases populate available features, so we need to handle that

I had considered moving everything down a level like install's op but I
decided to hold off for now just to minimize disruption, if nothing
else.