fix(set-version): Don't overwrite dependency updates
We load all manifests up front and and then iteratively update and write
them out.  The problem is between load and update/write, we might have
processed another manifest and updated the dependency table in the other
manifest, causing the last writer to win and updates to be lost.

Our tests haven't forced this situation yet.  We inherited this from
`upgrade` but since it only write to one at a time, it doesn't have the
last-writer-wins problem.

Fixes #702