Remove cargo-add
I was originally planning on keeping this around for a while but in
trying to sync `cargo-edit` lib code back from cargo will require
significant changes to the `add` command and it just doesn't seem worht
it.  Instead, we are letting people know what version they can install
to access the old command.