fix(upgrade): `--to-lockfile` changes all sources
With path and git dependencies, you can specify a version for use with
the registry.  If its incompatible, I believe the resolver will fail, so
we don't need to handle that case.  That just leaves `--to-lockfile`.
This just means post-poning the check for the registry to be in the
non-lockfile case.

Fixes #426