fix(upgrade): use dummy git repositories for tests
Fixes #764, in which upgrade tests were failing due to a more recent version of
serde in the git repository.

This is resolved by creating a dummy serde repository and referencing it by path
instead of by URL. This should hopefully preserve most of the testing surface
area of the original test.