fix(upgrade): Always upgrade compatible dependencies
As part of feedback on `cargo-upgrade` being merged into `cargo`, we are
looking into how we can make `upgrade` fit into `update` or be a direct
replacement for `update`.  In doing so, the "skip compatible by default"
workflow doesn't quite fit the same.  Similarly, manifests, in theory,
will always match the lock file, so there will be less of a reason to
upgrade the manifests only.

We'll need to re-open #552 and #297 until we figure out how we want to
handle this.