feat(upgrade): Make incompatible opt-in
On the path to upstreaming into cargo, this is part of an experiment
that asks "what if we started from scratch, what
would `cargo update` look like?".  Besides getting us to think outside
the box, I hope that we can deprecate `cargo update` and replace it with
`cargo upgrade`.  To build up a case to justify the ecosystem churn, we
need to get run time with the expected behavior to see how it works.

This is disruptive to existing cargo-edit users though.  So why
move away from the traditional behavior?

Reason 1: The new model for `cargo upgrade` is "what if `cargo update` updated the
manifest to match the lock file".  `cargo update` only deals with
compatible upgrades and it would be disruptive to that common workflow
to deal with incompatible upgrades.

Reason 2: There are several workflows we are trying to cover that
involve needing to upgrade only a subset of dependencies.  Before, we
had been working under the assumption of incompatible-by-default which
required ways to opt-out which have been a bit awkward.
Compatible-by-default provides a simple way of supporting most documented
workflows
- Application authors who want the latest features and bug fixes can get
  it with a simple `cargo upgrade -i`
- Users who want to defer breaking changes can do so by leaving off
  `--incompatible` or having it but using it with `--exclude`
- For maintainers who want isolated PRs, so long as someone has done
  `cargo upgrade`, merged it, and then done `cargo upgrade
  --incompatible`, they'll get that.

The workflows we are underserving still are:
- Library authors who want to keep their compatible versions low will
  need to go through awkwards steps of `cargo upgrade --dry-run` and
  then re-running with `-p` for each incompatible dependency
- Checking for upgrades is easy but checking for only incomptible
  upgrades will be a pain

Our alteratives to these haven't been great though
- Provide `--no-compatible` which feels out of place and awkward to tack
  on
- Make the flags `--compatible=<true|false>`, etc which would be
  cumbersome to specify.  We could make them `true` if no value is
  specified.  This makes things feel more consistent but its still
  cumbersome to have incompatible-only upgrades (`--compatible=false`
  instead of `--no-compatible`).  This would fit in with
  `--recursive=<true|false>`.
- Make `--compatible` the default if no flags are specified but
  otherwise make the flags additive (`cargo upgrade` would upgade
  compatible but `cargo upgrade --incompatible` wouldn't and would
  require `cargo upgrade --incompatible --compatible`).  This feels like
  it'd be too confusing
- Make `--compatible` the default but make all flags mutually exclusive.
  This makes the specifying-a-specific case short but requires running
  2-3 times to cover multiple cases.  Road humps in the users workflow
  like this are frustrating.